feat: added sending email messages function

+++ NewPoint/Repositories/ICodeRepository.cs
@@ -2,6 +2,6 @@
 
 public interface ICodeRepository
 {
-    Task AddEmailCode(int userId, string email, string code, int lifeTime = 120);
-    Task<bool> VerifyEmailCode(int userId, string email, string code);
+    Task AddEmailCode(long userId, string email, string code, int lifeTime = 120);
+    Task<bool> VerifyEmailCode(long userId, string email, string code);
 }
